Rebase and update helpers to utils
